Move NewRecipe, revert Title and SearchBar, update font to Inter and adjust form input types
