+++ src/Components/SearchBar.jsx
@@ -1,4 +1,3 @@
-import React, { useState } from "react";
 import { FaSearch } from "react-icons/fa";
 import "./SearchBar.css";
