--- src/Components/Title.css
@@ -1,23 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-.title-container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  width: 100%;
-  text-align: center;
-}
-
-.title-container h1 {
-  font-size: 4rem;
-  padding-bottom: 10px;
-}
-
-.title-container p {
-  font-size: 1.2rem;
-  padding-bottom: 50px;
-}
