+++ src/Components/Title.jsx
@@ -1,12 +1,8 @@
-import "./Title.css";
-
-export default function Title() {
-  return (
-    <>
-      <div className="title-container">
-        <h1>Bite Builder</h1>
-        <p>A tool to help you use up those leftover ingredients</p>
-      </div>
-    </>
-  );
+export default function Title(){
+  return(
+    <div>
+      <h1>Recipe Finder</h1>
+      <p>A tool to help you use up those leftover ingredients</p>
+    </div>
+  )
 }
