+++ src/pages/NewRecipe.css
@@ -1,3 +1,7 @@
+* {
+  font-family: "Inter", sans-serif;
+}
+
 .new-recipe-container {
   max-width: 1200px;
   margin: auto;
@@ -51,7 +55,7 @@ textarea:focus {
 
 .ingredients-container,
 .right-side {
-  flex: 1; 
+  flex: 1;
 }
 
 .right-side {
@@ -67,7 +71,7 @@ textarea:focus {
   flex-direction: column;
   align-items: center;
   justify-content: center;
-  position: relative; 
+  position: relative;
 }
 
 .image-placeholder,
@@ -81,7 +85,7 @@ textarea:focus {
   justify-content: center;
   font-size: 20px;
   color: black;
-  cursor: pointer; 
+  cursor: pointer;
   text-align: center;
 }
 
@@ -103,10 +107,10 @@ textarea:focus {
 }
 
 .detail-item {
-  display: flex; 
-  flex-direction: column; 
-  justify-content: center; 
-  align-items: center; 
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
   width: 30%;
   text-align: center;
 }
@@ -124,13 +128,13 @@ span {
   padding: 10px;
   margin-bottom: 10px;
   box-sizing: border-box;
-  width: 100%; 
+  width: 100%;
 }
 
 .total-time {
   padding: 10px;
-  width: 100%; 
-  text-align: center; 
+  width: 100%;
+  text-align: center;
 }
 
 button {
